refactor schema export integration api
